pyproject.toml
[tool.poetry]
name = "thought-agents"
version = "0.2.0"
description = "package for thoughtful agents partipating in stimulating, interactive conversations in a intra- & intro- spective manner"
authors = ["supermomo668 <mymomo119966.mm@gmail.com>"]
license = "Apache2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain-google-genai = "1.0.7"
openai = "1.35.13"
pandas = "2.2.2"
pillow = "10.4.0"
pydantic = "2.8.2"
python-dotenv = "1.0.1"
tqdm = "4.66.2"
duckduckgo-search = "6.1.12"
pyautogen = "0.2.31"
beartype = "0.18.5"
langchain = "0.2.6"
google-generativeai = "^0.7.2"
vertexai = "^1.49.0"

[tool.poetry.group.dev.dependencies]
hydra-core = "^1.3.2"
pytest = "^6.2.4"
omegaconf = "^2.1.1"
# Add the most common utilities you need for testing
pytest-cov = "^2.12.1"
pytest-mock = "^3.6.1"
tox = "^3.24.1"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "test-pypi"
url = "https://test.pypi.org/simple/"
priority = "explicit"